[analysis_server] Extract computing CodeActions from the textDocument/codeAction handler
In an upcoming change, we will need to be able to compute code actions from a command handler (to handle applying a code action as a result of a command call).

This change just extracts the main computation of code actions out of the handler into a new class that can be reused there.

This change is largely just a cut/paste, but I did extract a number of inline functions to be class-level (the variables they required are now at the class level) to reduce some nesting.

Change-Id: I438cfc07edfef9fa6d98c8e29c183b30abcf82d4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/427220
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>