Fine. Put declared fields into manifest.
Because there is separate getField(name).

And we remove the constant initializer from the getter ID, so if
the code uses only the type of a constant, probably quite often for
the analyzer, we don't re-analyze when the initializer changes,
because we depend only on the getter (type).

Change-Id: I585d45d4b392772842723e20e35e87ba2a88ee47
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/427600
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>