[analysis_server] Enable textDocument/codeActions and related refactor commands for LSP-over-Legacy
This moves the codeAction handler and the refactor commands to be shared (and therefor work for either base server kind). It also adds LSP-over-Legacy test classes that use the shared mixins so that all relevant tests run in both modes.

It does _not_ yet enable this functionality for DTD (though this is a requirement to do so in a way that will work for either IDE/server).

Change-Id: I790b98dafc59e83e72b399c9b88b94003a102f66
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/428322
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>