[ddc] Update source mapping for static invocations
when they have been rewritten to include hot reload checks. Appears to
fix the issue detected in the dwds test failure.

Issue: https://github.com/dart-lang/webdev/issues/2617.

The new mapping ensures breakpoints applied to the call site in the Dart
source are the call site instead of the the generation check.

Allows the tooling to set a breakpoint on the static method call site,
and then when paused, a step-into will action will advance the program
into the static method.

In the future we will need to reevaluate how we generate code where a
single branch of execution in Dart source produces multiple branches in
JavaScript. I believe right now the tooling will not support creating
breakpoints in both branches even when they both have a mapping to the
same Dart source location.

Change-Id: I1ea0548768d8213dc7eefc332814d3bda7c81535
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/427567
Reviewed-by: Nate Biggs <natebiggs@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>
Reviewed-by: Srujan Gaddam <srujzs@google.com>