[analyzer] Dot shorthands: Avoid crashing with prefix and postfix operators.
When we evaluate the `target` for this `IndexExpression`, we don't cache the  dot shorthand context which causes the analyzer to crash. Added a context push of the `target` (whose context is `UnknownType`) which allows the analyzer to error out without crashing.

Passing co19 tests and unit tests.

Bug: https://github.com/dart-lang/sdk/issues/59835
Change-Id: I6a19b3e92ed803a51d0cfd320df886a26a36b33d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/426289
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>