[dart2wasm] Remove dynamic module hacks in dispatch table
The dispatch table was so far getting rows for selectors even though
we don't need to call those via dispatch table:

* if there's only 0 targets in the module then it's unreachable
* if there's only 1 targets in the module then we can call it directly

Only if we have >1 targets in the module do we need a dispatch
table entry for the selector.

So we update the dynamic module updatable function mechanism to handle
the

   * unreachable
   * direct call
   * dispatch table call

cases just like we do in the normal code generator.

Change-Id: I7144771d61fefa09c215e28dd498a4f888c73d22
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/428981
Reviewed-by: Nate Biggs <natebiggs@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>