Add experiment keys to experiment-related diagnostics
In a recent discussion we decided to add an `experiment` key to the
entries in the analyzer's `messages.yaml` file with the same semantics
as they have in the front-end messages file. This is an initial attempt
to add them.

I'm reasonably confident that all of the markes diagnostics are related
to the specified experiment, but there might well be others. I'm happy
to update this CL to capture those as well, but we can always add the
map entry later as we discover diagnostics that were missed.

Change-Id: I31ec94bae905828e64a1769fdf905ef7a19a8d11
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/430120
Reviewed-by: Kallen Tu <kallentu@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>