[dart2wasm] Some refactorings of the compiler
a) Usage of `ClassInfo.*`

We have an abstraction layer that translates dart types
to wasm types - which is the `Translator` object (it
has e.g. `Translator.translateType`).

=> We remove direct use of `ClassInfo.*` in most places
=> Instead of exposing `ClassInfo` for specific types
   (e.g. top type), we expose the wasm representation type

b) Type of `ClassInfo.repr`

Currently `ClassInfo.repr` is itself a `ClassInfo`, which is
rather confusing. We use `ClassInfo.repr` to determine how to
translate `InterfaceType`s of that class to the wasm type system.

=> We make `Class.repr` be an actual wasm representation type
   (we make it a `w.RefType`)

c) We create a macro assembler `loadClassId` method so we have
   one place where we load the class id and we verify in
   assertions that the receiver type is `ref #Top`.

Change-Id: Ie24008a98131dae34948406a1f5aec881499c1f0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/431080
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>