Elements. Make FragmentOrMember implement Fragment.
I few references to `Fragment` were converted to `FragmentImpl` because
it was a convenient time to do.

We added throwing implementations of `Fragment` methods to
`PrefixElementImpl` and `Member` temporarily. We remove
`PrefixElementImpl` completely in the next CL, and eventually `Member`
also will stop implementing `Fragment`, and will implement only
`Element`. But for now it would be a too big change to me to do at once.

Change-Id: I9cf9034a9c45e9baee6ab7421d7ba4d50b6c45c5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/431622
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>